fix(calculator): guard two buffer sizes, and hold the line with CI
Calculator was the largest project here and had the least CI: no
warnings gate, no coverage. It also built under -Wall -Wextra only,
where the other two use five flags.

Turning the other three on found seven conversion warnings. Five are
benign -- a length checked against sizeof on the line above, then passed
to strncpy -- and are now explicit casts, placed after the guard that
makes them safe rather than instead of one.

Two were not benign. popString and peekString take the output buffer's
size as an int and never checked it:

    strncpy(out, s->items[s->top--], outSize - 1);
    out[outSize - 1] = '\0';

At outSize = 0 that length is -1, which strncpy reads as a size_t: it
becomes SIZE_MAX and the copy runs until it happens upon a NUL, and the
next line writes out[-1]. Every caller today passes sizeof(buffer), so
this was a hazard rather than a live defect -- but refusing a size it
cannot honour is the callee's job, and the compiler cannot tell the two
guarded cases from these two unguarded ones. That is the argument for
the flag.

Added to CI: a -Werror job under all five flags, and a coverage job.
Coverage needed building first -- a BUILD=coverage mode and a gcov
script, modelled on DigitalClock's. It reports 81.49% across 1972 lines,
enforced at 78%, excluding main.c, which is the interactive loop the
suite does not drive and would otherwise show as dead code.

`coverage` is declared .PHONY, which is not a formality: the script
writes a directory of that name, and without it Make saw the directory,
decided the target was up to date, and did nothing -- silently, and
including in CI. Caught by running it twice.

486 tests pass under -Werror, ASan, UBSan and CMake; the Qt GUI builds
clean under the new flags too.

+++ Calculator/Src/main.c
@@ -201,7 +201,8 @@ int main(void)
 
             if (equal != NULL)
             {
-                int len = equal - infix;
+                /* infix is a bounded buffer, so the difference fits. */
+                int len = (int)(equal - infix);
 
                 if (len <= 0)
                 {
@@ -215,7 +216,8 @@ int main(void)
                     break;
                 }
 
-                strncpy(variableName, infix, len);
+                /* len is > 0 and < sizeof(variableName) by the checks above. */
+                strncpy(variableName, infix, (size_t)len);
                 variableName[len] = '\0';
 
                 char *start = variableName;
